Extract Data From Tweeter And Save In Azure SQL Using Azure ServerLess/requirements.txt
azure-functions
pandas
twython
pandas
pyodbc
sqlalchemy